LayoutTests/imported/w3c/web-platform-tests/resource-timing/rt-initiatorType-media-expected.txt

Resource Timing: Element localName initiatorType


FAIL Type: video - <video src> assert_equals: initiatorType should be 'video' expected "video" but got "other"
PASS Timing: video - <video src> 
FAIL Type: audio - <audio src> assert_equals: name should be the url expected "http://localhost:8800/resource-timing/resources/data.json?audio&0.7734679755996403" but got "http://localhost:8800/resource-timing/resources/data.json?video&0.5872075656484804"
PASS Timing: audio - <audio src> 
